product list display update details & add product code login validation

+++ app/src/main/java/com/basicneedscorporation/basicneeds/ui/manageorder/cart/CartActivity.kt
@@ -69,7 +69,7 @@ class CartActivity : BaseActivity(), IAdapterOnClick {
     }
 
     override fun onClick(item: Any, position: Int) {
-        if (item is Int) {
+        if (item is Float) {
             textViewOrderAmount.text = "₹ " + item.toString()
         }
     }
